version 0.5: fix blurdle after breakage
data-pagelet attribute we were selecting on (for each feed post) is
gone

instead, look for a very precise (and fragile) class attribute on the
div one level down. this defines a feed post.

continue grep'ing for wordle as before.

but now once we filter, we need to grab the parent <div> tag, before
removing.

update both code and README (which previously just noted the breakage)